+++ src/main/java/com/hanshin/supernova/config/web/WebConfig.java
@@ -2,7 +2,7 @@ package com.hanshin.supernova.config.web;
 
 import com.hanshin.supernova.config.resolver.UserArgumentResolver;
 import com.hanshin.supernova.redis.community_stat.interceptor.SingleVisitInterceptor;
-import com.hanshin.supernova.redis.visit.QuestionVisitInterceptor;
+import com.hanshin.supernova.redis.question_visit.interceptor.QuestionVisitInterceptor;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.CacheControl;
